Btrfs: fix slot count logic in space info ioctl
Patchwork-id: 33192
O-Subject: [RHEL6.1 PATCH] Btrfs: fix slot count logic in space info ioctl
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Petr Matousek <pmatouse@redhat.com>
RH-Acked-by: Eugene Teo <eugene@redhat.com>

This is for bz 663749 and is from Dan Rosenberg.  The ioctl takes a u64 for the
number of slots userspace has for it's space info.  The problem is we're using
an int to store that in the ioctl, so that means we can end up with a negative
value for allocating the destination buffer.  Fix this by making sure we use u64
the the local slot count.  Also we need to make sure we don't copy in more space
info's than we have slots for, so keep track of how many slots have been used
and break if we use them all up.

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>