qeth: wrong MAC-address displayed in error message
Patchwork-id: 33282
O-Subject: [RHEL6.1 PATCH 1/2] [s390x] qeth: wrong MAC-address displayed in
	error message
Bugzilla: 675211
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Description
-----------
Setting of a MAC-address might fail because an already used
MAC-address is to bet set or because of authorization problems.
In those cases qeth issues a message, but the mentioned
MAC-address is not the new MAC-address to be set, but the actual
MAC-address.

Choose the new MAC-address to be set for the error messages.

Bugzilla
--------
BZ 675211
https://bugzilla.redhat.com/show_bug.cgi?id=675211

Upstream status of the patch
----------------------------
http://git.kernel.org/?p=linux/kernel/git/davem/net-2.6.git;a=commitdiff;h=221c17fe87033aa154df68679b437c83d835c28

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>