dasd: prevent panic with unresumed devices
Patchwork-id: 33283
O-Subject: [RHEL6.1 PATCH 2/2] [s390x] dasd: prevent panic with unresumed devices
Bugzilla: 676720
RH-Acked-by: Thomas Graf <tgraf@redhat.com>

Description
-----------
A kernel panic might occur when an unresumed device is set offline.

During suspend we deregister the device from alias handling and
on resume we register it again to assure that we have the correct
alias device combination set.
During offline processing we also deregister devices from alias
handling an so we might refer to structures that are not available
at present.

Check if the appropriate lcu structure is available. If it is not,
skip the deregistration process because the device is already removed.

Bugzilla
--------
BZ 676720
https://bugzilla.redhat.com/show_bug.cgi?id=676720

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.38
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=f602f6d694a99a0141c066c8f0b360a0b3c16915

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>