ixgbe: fix panic due to uninitialized pointer
Patchwork-id: 33290
O-Subject: [PATCH RHEL6.1] ixgbe: fix panic due to uninitialized pointer
Bugzilla: 676875
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Systems containing an 82598 were panicing on boot.  It turns out
hw->mac.ops.setup_sfp is only set for 82599, so one should check to be
sure that pointer is set before continuing in
ixgbe_sfp_config_module_task.

The upstream driver also has this problem, so I posted a patch to
netdev to address it upstream.

http://marc.info/?l=linux-netdev&m=129746077110725&w=2

Intel has agreed to test it and post it for inclusion in their next
round of ixgbe changes.

Verified by me that this patch resolves RHBZ 676875.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>