sha-s390: Reset index after processing partial block
Patchwork-id: 33324
O-Subject: [RHEL6 PATCH] crypto: sha-s390 - Reset index after processing
	partial block
Bugzilla: 626515
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

RHEL6 Bugzilla #626515

This patch fixes a bug in S390's implementation of SHA that not
only can produce incorrect hashes, but that also has the potential
to corrupt memory.

crypto: sha-s390 - Reset index after processing partial block

The partial block handling in sha-s390 is broken when we get a
partial block that is followed by an update which fills it with
bytes left-over.  Instead of storing the newly left-over bytes
at the start of the buffer, it will be stored immediately after
the previous partial block.

This patch fixes this by resetting the index pointer.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

Thanks,

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>