hpsa: sanitize max commands
Patchwork-id: 33352
O-Subject: [RHEL6.1 PATCH 031/112] hpsa: sanitize max commands
Bugzilla: 630060
RH-Acked-by: Rob Evers <revers@redhat.com>

Some controllers might try to tell us they support 0 commands
in performant mode.  This is a lie told by buggy firmware.
We have to be wary of this lest we try to allocate a negative
number of command blocks, which will be treated as unsigned,
and get an out of memory condition.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>