hpsa: disable doorbell reset on reset_devices
Patchwork-id: 33359
O-Subject: [RHEL6.1 PATCH 036/112] hpsa: disable doorbell reset on reset_devices
Bugzilla: 630060
RH-Acked-by: Rob Evers <revers@redhat.com>

The doorbell reset initially appears to work correctly,
but on some Smart Arrays, it causes a subsequent controller
lockup, and kdump can end up leaving the root filesystem
in an unbootable state.  For this reason, until the
problem is fixed, or isolated enough to be avoided,
the doorbell reset not be used at all.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>