add sysfs knob for turning off disk entropy contributions
Patchwork-id: 33451
O-Subject: [RHEL 6 PATCH] block: add sysfs knob for turning off disk entropy
	contributions
Bugzilla: 677447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

Hi,

This is a simple backport of the following upstream commit:

  commit e2e1a148bc45855816ae6b4692ce29d0020fa22e
  Author: Jens Axboe <jaxboe@fusionio.com>
  Date:   Wed Jun 9 10:42:09 2010 +0200

    block: add sysfs knob for turning off disk entropy contributions

    There are two reasons for doing this:

    - On SSD disks, the completion times aren't as random as they
      are for rotational drives. So it's questionable whether they
      should contribute to the random pool in the first place.

    - Calling add_disk_randomness() has a lot of overhead.

    This adds /sys/block/<dev>/queue/add_random that will allow you to
    switch off on a per-device basis. The default setting is on, so there
    should be no functional changes from this patch.

    Signed-off-by: Jens Axboe <jaxboe@fusionio.com>

The security team brought this up due to FIPS certification work.*  The
patch is straight-forward, and as the changelog notes, it doesn't modify
the default behaviour.  I booted a brew-built kernel with this patch to
ensure that the add_random file appeared in /sys/block/sdX/queue, and
that it could be toggled.  This resolves bug 677447.  Comments, as
always, are appreciated.

Cheers,
Jeff

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>