ipv6: Implement Any-IP support for IPv6
Patchwork-id: 33468
O-Subject: Re: [RHEL 6.1 PATCH] ipv6: Implement Any-IP support for IPv6
Bugzilla: 591335
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>

I apologize the original patch was corrupt.

This is a backport of upstream commit:
ab79ad14a2d51e95f0ac3cef7cd116a57089ba82

Provides AnyIP capability, which is the ability to receive packets and
establish incoming connections on IPs we have not explicitly configured
on the machine.

An example use case is to configure a machine to accept all incoming
traffic on eth0, and leave the policy of whether traffic for a given IP
should be delivered to the machine up to the load balancer. Integral
functionality in support of IPv6 TPROXY.

Please note that this is a follow-up patch to the patch series already
posted in the Bugzilla which is required for IPv6 TPROXY to work properly.

Bugzilla:
---------
https://bugzilla.redhat.com/show_bug.cgi?id=591335

Brew:
-----
https://brewweb.devel.redhat.com/taskinfo?taskID=3119462

Test Status:
------------
The customer was able to verify that this patch works.
Comments are in the Bugzilla.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>