cm: Fix crash in request handlers
Patchwork-id: 33608
O-Subject: [RHEL6.1 PATCH 1/2] rdma/cm: Fix crash in request handlers
Bugzilla: 676191
RH-Acked-by: Jay Fenlason <fenlason@redhat.com>

This is needed for the 6.1 kernel.  It addresses a valid use scenario
that can oops the kernel (and a malicious user could oops the kernel
rather quickly given the right conditions).  Since it can allow a remote
attacker (or local user, depending on your point of view) to oops a
machine, it carries a CVE number.  Bug for 6.1 is 676191.  There are two
patches requires to solve this bug, this is the first of them.

-------- Forwarded Message --------
From: Hefty, Sean <sean.hefty@intel.com>
To: linux-rdma@vger.kernel.org <linux-rdma@vger.kernel.org>, Roland
Dreier <roland@kernel.org>
Cc: Doug Ledford <dledford@redhat.com>
Subject: [PATCH 1/2] rdma/cm: Fix crash in request handlers
Date: Wed, 23 Feb 2011 08:11:32 -0800

Doug Ledford and RedHat reported a crash when running the rdma_cm on a real
time OS.  The crash has the following call trace:

    cm_process_work
       cma_req_handler
          cma_disable_callback
          rdma_create_id
             kzalloc
             init_completion
          cma_get_net_info
          cma_save_net_info
          cma_any_addr
             cma_zero_addr
          rdma_translate_ip
             rdma_copy_addr
          cma_acquire_dev
             rdma_addr_get_sgid
             ib_find_cached_gid
             cma_attach_to_dev
          ucma_event_handler
             kzalloc
             ib_copy_ah_attr_to_user
          cma_comp

[ preempted ]

    cma_write
        copy_from_user
        ucma_destroy_id
           copy_from_user
           _ucma_find_context
           ucma_put_ctx
           ucma_free_ctx
              rdma_destroy_id
                 cma_exch
                 cma_cancel_operation
                 rdma_node_get_transport

        rt_mutex_slowunlock
        bad_area_nosemaphore
        oops_enter

They were able to reproduce the crash multiple times with the following
details:

    Crash seems to always happen on the:
            mutex_unlock(&conn_id->handler_mutex);
    as conn_id looks to have been freed during this code path.

An examination of the code shows that a race exists in the request handlers.
When a new connection request is received, the rdma_cm allocates a new
connection identifier.  This identifier has a single reference count on it.
If a user calls rdma_destroy_id() from another thread after receiving a
callback, rdma_destroy_id will proceed to destroy the id and free
the associated memory.  However, the request handlers may still be in
the process of running.  When control returns to the request handlers,
they can attempt to access the newly created identifiers.

Fix this by holding a reference on the newly created rdma_cm_id until
the request handler is through accessing it.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>