cm: Bump reference count on cm_id before invoking callback
Patchwork-id: 33609
O-Subject: [RHEL6.1 PATCH 2/2] ib/cm: Bump reference count on cm_id before
	invoking callback
Bugzilla: 676191
RH-Acked-by: Jay Fenlason <fenlason@redhat.com>

This is needed for the 6.1 kernel.  It addresses a valid use scenario
that can oops the kernel (and a malicious user could oops the kernel
rather quickly given the right conditions).  Since it can allow a remote
attacker (or local user, depending on your point of view) to oops a
machine, it carries a CVE number.  Bug for 6.1 is 676191.  There are two
patches requires to solve this bug, this is the second of them.

-------- Forwarded Message --------
From: Hefty, Sean <sean.hefty@intel.com>
To: linux-rdma@vger.kernel.org <linux-rdma@vger.kernel.org>, Roland
Dreier <roland@kernel.org>
Cc: Doug Ledford <dledford@redhat.com>
Subject: [PATCH 2/2] ib/cm: Bump reference count on cm_id before
invoking callback
Date: Wed, 23 Feb 2011 08:17:40 -0800

When processing a SIDR REQ, the ib_cm allocates a new cm_id.  The refcount
of the cm_id is initialized to 1.  However, cm_process_work will
decrement the refcount after invoking all callbacks.  The result is that
the cm_id will end up with refcount set to 0 by the end of the sidr req
handler.

If a user tries to destroy the cm_id, the destruction will proceed, under
the incorrect assumption that no other threads are referencing the cm_id.
This can lead to a crash when the cm callback thread tries to access
the cm_id.

This problem was noticed as part of a larger investigation with kernel
crashes in the rdma_cm when running on a real time OS.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>