fix rx queue refcounting
Patchwork-id: 33611
O-Subject: Re: [RHEL6.1 PATCH] fix rx queue refcounting (bz 677786)
Bugzilla: 677786
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Amerigo Wang <amwang@redhat.com>

	My initial patch did in fact fix a refcounting problem in the origional
implementation of rps.  Unfortunately, it didn't actually fix the problem in
that, when the refcount equaled zero, we could still be using the dev->_rx array
while it was being freed.  My initial backport worked because I failed to remove
the initial refcount set that amerigo noted.  When I removed those, the crash
returned.  In summary, its still possible to have frames comming in for a device
as we're dropping references to dev->_rx in rx_queue_release, and we have no
protection around access to that pointer, so we blow up if the timing is right
(which it seems to be consistently in the LTC containers test suite).  Anywho,
this is a backport of upstream commit fe8222406c8277a21172479d3a8283d31c209028,
in which tom herbert simplified the ref counting, and in so doing, gated the
free of dev->_rx on the dropping of the last reference to the net device as a
whole, which guaranteed there are no rx path users prior to its release.  I just
tested this out several times and it solved bz 677786.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>