fcoe: Fix module reference count for vports
Patchwork-id: 33651
O-Subject: [RHEL6.1 PATCH 02/35] fcoe: Fix module reference count for vports
Bugzilla: 635898 668114
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Robert Love <robert.w.love@intel.com>

Merged in scsi-misc
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=7287fb9114096503eddfffb7b2ed691c809a6106

vports are not grabbing module references but are
releasing them. This causes the module reference count
to decrement too many times and it wraps around past 0.

The solution is to do a module_put() in
fcoe_interface_release() so that the reference is only
released when the fcoe_interface is released. There is a
one-to-one relationship between the N_Port and the
fcoe_interface, so the module reference will only be
dropped when the N_Port is destroyed

To create symetry in the code this patch moves the
try_module_get() call into fcoe_interface_create(). This
means that only the N_Port will grab a reference to the
module when its corresponding fcoe_interface is created.

This patch also makes it so that the fcoe_interface_create()
routine encodes any error codes in the fcoe_interface
pointer returned. This way its caller, fcoe_create(), can
return an accurate error code.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>