libfc: add hook to notify providers of local port changes
Patchwork-id: 33654
O-Subject: [RHEL6.1 PATCH 08/35] libfc: add hook to notify providers of local
	port changes
Bugzilla: 635898 668114
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Joe Eykholt <jeykholt@cisco.com>

Merged in scsi-misc
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=70d53b046a6221e3ceb3bd8eaa807ef6a1c53762

When an SCST provider is registered, it needs to know what
local ports are available for configuration as targets.

Add a notifier chain that is invoked when any local port
that is added or deleted.

Maintain a global list of local ports and add an
interator function that calls a given function for
every existing local port.  This is used when first
loading a provider.

Signed-off-by: Joe Eykholt <jeykholt@cisco.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>