fcoe: prepare fcoe for using fcoe transport
Patchwork-id: 33658
O-Subject: [RHEL6.1 PATCH 17/35] fcoe: prepare fcoe for using fcoe transport
Bugzilla: 635898 668114
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Yi Zou <yi.zou@intel.com>

Merged in scsi-misc
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=commit;h=8ca86f84dd5fc881012b2940ce4eab1fa57680e5

Prepare the fcoe to convert it to use the newly added fcoe transport, making
it as the default fcoe transport provider for libfcoe. This patch is to rename
some of the variables to avoid any confusing names later as now there are
several transports in the same file.

Signed-off-by: Yi Zou <yi.zou@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>