libfc: revert patch to fix exchange being deleted when the abort itself is timed out
Patchwork-id: 33669
O-Subject: [RHEL6.1 PATCH 23/35] libfc: revert patch to fix exchange being
	deleted when the abort itself is timed out
Bugzilla: 635898 668114
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This fixes a regression added with
commit 15186ce42f79936dddd04e20a2479698b9f9a336
[fcoe] libfc: fix exchange being deleted when the abort itself is timed out

This has been sent here
https://lists.open-fcoe.org/pipermail/devel/2011-February/011077.html

>From Neerav Parikh.

When abort for an exchange timed out it didn't release the reference to
the exchange resulting in a memory leak.

After discussion with the author of the patch (CC) that introduced this
bug it was suggested to revert that patch.

This reverts commit ea3e2e72eeb3e8a9440a5da965914f9b12088626.

Signed-off by: Neerav Parikh <Neerav.Parikh at intel.com>
Tested-by: Ross Brattain <ross.b.brattain at intel.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>