fix pgd_lock deadlock
Patchwork-id: 33688
O-Subject: [PATCH RHEL6.1] fix pgd_lock deadlock #2
Bugzilla: 671477
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

This is a new approach. Waiting upstream to comment on it to be sure
it's enough. If it's enough the removal of _irqsave isn't needed
anymore for RHEL (I'll push it later upstream as a cleanup).

https://bugzilla.redhat.com/show_bug.cgi?id=671477
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3143355

===
Subject: xen: stop taking the page_table_lock with irq disabled

From: Andrea Arcangeli <aarcange@redhat.com>

It's forbidden to take the page_table_lock with the irq disabled or if there's
contention the IPIs (for tlb flushes) sent with the page_table_lock held will
never run leading to a deadlock.

Only Xen needs the page_table_lock and Xen won't need IPI TLB flushes hence the
deadlock doesn't exist for Xen.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>