compaction: Avoid a potential deadlock due to lock_page() during direct compaction
Patchwork-id: 33692
O-Subject: [PATCH RHEL6.1 03/22] mm: compaction: Avoid a potential deadlock due
	to lock_page() during direct compaction
Bugzilla: 674147
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

From: Mel Gorman <mel@csn.ul.ie>

Hugh Dickins reported that two instances of cp were locking up when
running on ppc64 in a memory constrained environment. It also affects
x86-64 but was harder to reproduce. The deadlock was related to readahead
pages. When reading ahead, the pages are added locked to the LRU and queued
for IO. The process is also inserting pages into the page cache and so is
calling radix_preload and entering the page allocator. When SLUB is used,
this can result in direct compaction finding the page that was just added
to the LRU but still locked by the current process leading to deadlock.

This patch avoids locking pages in the direct compaction patch because
we cannot be certain the current process is not holding the lock. To do
this, PF_MEMALLOC is set for compaction. Compaction should not be
re-entering the page allocator and so will not breach watermarks through
the use of ALLOC_NO_WATERMARKS.

Reported-by: Hugh Dickins <hughd@google.com>
Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>