thp: Use correct numa policy node for transparent hugepages
Patchwork-id: 33708
O-Subject: [PATCH RHEL6.1 17/22] thp: Use correct numa policy node for
	transparent hugepages
Bugzilla: 674147
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: Andi Kleen <ak@linux.intel.com>

Pass down the correct node for a transparent hugepage allocation.
Most callers continue to use the current node, however the hugepaged
daemon now uses the previous node of the first to be collapsed page
instead. This ensures that khugepaged does not mess up local memory
for an existing process which uses local policy.

The choice of node is somewhat primitive currently: it just
uses the node of the first page in the pmd range. An alternative
would be to look at multiple pages and use the most popular
node. I used the simplest variant for now which should work
well enough for the case of all pages being on the same node.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>