remove execute bit from perf scripts
Patchwork-id: 33805
O-Subject: [RHEL6.1 PATCH] remove execute bit from perf scripts
Bugzilla: 676692
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>

Hi,

'perf' comes with a number of .py and .pl scripts which are marked as
executable by the filesystem. However, these scripts are not directly exectued
but instead executed 'indirectly' through perf launching the corresponding
script interpreter. Thus, not all of the .py, .pl scripts begin with a:
#!/usr/bin/perl or #!/usr/bin/python line. This confuses, 'find-requires'
a causes an error when building the perf pkg:

Finding  Requires: /usr/lib/rpm/redhat/find-requires kernel
error: line 8726: Dependency tokens must begin with alpha-numeric, '_' or '/':
error: Failed to find Requires:

Although this is arguably a 'find-requires' bug, we can work around this issue
in the kernel by simply not marking the .pl, and .py files as executable. This
doesn't appear to change how perf behaves in any way, and fixes this annoying
msg, at least until 'find-requires' is updated.

This is the issue that Steve Dickson reported, which is worked around by
downgrading, the file pkg to '-5'.

Fixes bz #676692.

thanks,

-Jason

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>