tape: deadlock on global work queue
Patchwork-id: 33824
O-Subject: [RHEL6.1 PATCH 3/3] [s390x] tape: deadlock on global work queue
Bugzilla: 681306

Description
-----------
The system might hang when using the tape device driver.

The 34xx and 3590 tape driver use the system work queue to
defer work from the interrupt function to process context.
The same work queue is used to retry a command if it failed
with a deferred condition code. Therefore the work function
may not wait for the end of a command, otherwise a deadlock
occurs.

Move the post processing of the tape commands started via the
global work queue to the request callback function.

Bugzilla
--------
BZ 681306
https://bugzilla.redhat.com/show_bug.cgi?id=681306

Upstream status of the patch
----------------------------
The patch has been sent upstream for 2.6.38-rc7
https://lkml.org/lkml/2011/3/7/83

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>