radix: don't tag the root if we didn't tag within our range
Patchwork-id: 33859
O-Subject: [RHEL6.1 PATCH] radix: don't tag the root if we didn't tag within
	our range
Bugzilla: 681439
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

Hello,

This fix is in reference to bz 681439 and is a backport of the commit

ac15ee691fe84cb46cbd2497ddcb10e246f7ee47

Basically we can end up in an infinite loop because we set the root node of the
radix tree with what we are trying to tag unconditionally, when we need to only
do it if we actually found something in our range to tag.  The commit has a big
long explanation for this.  We hit the problem described in the commit and this
patch fixes the problem.  Thanks,

Josef

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>