allow kernel-debuginfo-common to be biarch installable
Patchwork-id: 33886
O-Subject: [RHEL 6.1 PATCH] allow kernel-debuginfo-common to be biarch
	installable - again
Bugzilla: 682012
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Roland McGrath <roland@redhat.com>

Hi,

In 6.1, we've moved the perf binary into the 'perf' pkg itself. However,
this has made the kernel-debuginfo-common pkg for x86 and x86_64
conflict b/c perf.debug is common to both.

Fix this by introducing a new perf-debuginfo pkg. The perf debugging
information is not part of the kernel proper anyway. With this change, I
can now succesfully install the kernel-debuginfo-common pkg for both x86
and x86_64 on the same system.

Thanks to Aris for pointing out this issue, and to Roland for the
implementation, which he has also introduced into fedora.

Resolves bz#682012.

thanks,

-Jason

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>