gfs2: quota allows exceeding hard limit
Patchwork-id: 33913
O-Subject: [RHEL6.1][GFS2 PATCH] - Bug 675944 - gfs2 quota allows exceeding
	hard limit
Bugzilla: 675944
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

Immediately after being synced to disk, cached quotas are zeroed out and
a subsequent access of the cached quotas results in incorrect zero values.
This meant that gfs2 assumed the actual usage to be the zero (or near-zero)
usage values it found in the cached quotas and comparison against
warn/limits never triggered a quota violation.

This patch adds a new flag QDF_REFRESH that is set after a sync so that the
cached quotas are forcefully refreshed from disk on a subsequent access on
seeing this flag set.

Resolves: rhbz#675944
Signed-off-by: Abhi Das <adas@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>