nv50: insert a delay before fb change to prevent display engine hang
Patchwork-id: 33924
O-Subject: [RHEL6.1 PATCH 1/3] drm/nv50: insert a delay before fb change to
	prevent display engine hang
Bugzilla: 658896
RH-Acked-by: Dave Airlie <airlied@redhat.com>

This is a re-posting of a patch originally written for rhbz#618225.  This
appears to have gotten lost during the DRM update for 6.1.

The display engine on this card appears to strongly dislike when we submit
multiple framebuffer-only display switches too soon after each other (as
in the dual-head case here, when switching to/from X and a VT), and gets
stuck in a state where it will no longer accept additional updates.

This problem has now been fixed in a more correct manner upstream, but is
far too intrusive a change to propose for 6.1 at this time.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>