hpsa: export resettable host attribute
Patchwork-id: 33950
O-Subject: [RHEL6.1 PATCH 2/3] hpsa: export resettable host attribute
Bugzilla: 682239
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>
RH-Acked-by: Amerigo Wang <amwang@redhat.com>

This attribute allows kexec-tools to know
whether the controller can honor the reset_devices kernel parameter
and actually reset the controller. For kdump to work properly it
is necessary that the reset_devices parameter be honored.  This
attribute enables kexec-tools to warn the user if they attempt to
designate a non-resettable controller as the dump device.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>