capabilities: do not special case calculation of pE during exec of init
Patchwork-id: 33959
O-Subject: [RHEL6.1 PATCH] capabilities: do not special case calculation of pE
	during exec of init
Bugzilla: 684008
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: James Morris <jmorris@redhat.com>

RHBZ: 684008
upstream posting: http://marc.info/?l=linux-security-module&m=129988519927341&w=2

The problem is that even if you drop capabilities before you exec init the
kernel has special logic and init will have all the capabilities.  This
patch allows us to drop capabilities before the exec of init and they stay
dropped.

This patch differs from upstream because I decided to include a command
line option to enable it.  Upstream I just dropped the special logic
altogether.  We believe that this is safe upstream but it has absolutely
no testing.  In order to be 100% confident that we cannot possibly have a
regression at this late period of time in 6.1 for the 99% of users who don't
care to bound the capabilities on their /sbin/init I thought it prudent.

upstream changelog:
When the global init task is exec'd we have special case logic to make sure
the pE is not reduced.  There is no reason for this.  If init wants to drop
it's pE is should be allowed to do so.  Remove this special logic.

Signed-off-by: Eric Paris <eparis@redhat.com>
Acked-by: Serge Hallyn <serge@hallyn.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>