nouveau: disable acceleration on NVA3/NVA5/NVA8/NVAF by default
Patchwork-id: 33995
O-Subject: [RHEL6.1 PATCH] nouveau: disable acceleration on NVA3/NVA5/NVA8/NVAF
	by default
Bugzilla: 684816
RH-Acked-by: Dave Airlie <airlied@redhat.com>

rhbz#684816

The below patch got lost during the DRM rebase, and will lead to these
lockups occuring for people in 6.1 if it's not reapplied.

----
There's an GPU lockup problem for which the cause is currently unknown
on these chipsets.

Until it's resolved, it's better to leave the user with a working system
without acceleration than to have random lockups.

With this patch, acceleration will be off by default if a known problem
chipset is detected, but can be re-enabled with nouveau.noaccel=0 on
the kernel commandline.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>