export SMBIOS version via sysfs
Patchwork-id: 34000
O-Subject: [RHEL6 PATCH BZ 684329 V2]: Export SMBIOS version via sysfs
Bugzilla: 684329
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Bob Picco <bpicco@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

This patch exports the SMBIOS version via sysfs.  It uses the existing DMI
code and exports the value via

/sys/devices/virtual/dmi/id/smbios_version

This patch is RHEL6 only.  I'm going to put together a patch to export
SMBIOS info upstream over the next few weeks and push that upstream.  The
patch will create a similar layout to the dmi sysfs entries.

This patch is need for the biosdevname & udev code to check the SMBIOS version.
The rules could be modified to call dmidecode for every device, however, that
is wasteful.

jcm raised a minor issue with the patch -- I am NOT checking for the existence
of memory @ 0xF0000.  The DMI code doesn't do it right now so I'm not worried
about it either ;) but if anyone strenously objects, I can add an export of
a e820 function and check the value ... but that is overkill IMO.

Successfully tested by me.

[root@amd-dinar-05 ~]# dmesg | grep SMBIOS
SMBIOS version 2.5 @ 0xF78E0
[root@amd-dinar-05 ~]# cat /sys/devices/virtual/dmi/id/smbios_version
2.5
[root@amd-dinar-05 ~]# dmidecode | grep SMBIOS
SMBIOS 2.5 present.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>