memcg: never OOM when charging huge pages, reinstated
Patchwork-id: 34002
O-Subject: [RHEL6.1 PATCH] memcg: never OOM when charging huge pages, reinstated
Bugzilla: 685161
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

Hi,

we had this problem fixed before (4de6dac [mm] memcontrol: never oom
when charging huge pages) but the fix was accidentally reverted
through an unrelated change in the area (51d4547 [mm] Out-of-memory
under memory cgroup can call both of oom-killer-for-memcg and
oom-killer-for-page-fault).

This is a regression from 6.0 to 6.1, so I propose this fix to get in
before 6.1 is released.

Instead of reverting to the arrangement we had before, I backported
the upstream fix for this problem instead.