btrfs: bring us up to date with .38
Patchwork-id: 34055
O-Subject: [RHEL6.1 PATCH] btrfs: bring us up to date with .38
Bugzilla: 684667
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Hello,

I'm sorry I wanted to avoid this but we have had too many cricital fixes come in
between .37 and .38 that I can't ignore.  This is for bz 684667.

Basically there is one big feature, which is LZO compression, and the rest of
this is bug fixes.  They include 2 corruption fixes and 3 different hangups that
are easy to invoke.  I'm just giving one big patch for the update, if you want
to review the individual patches I've included them in my git tree

git://git.engineering.redhat.com/users/jwhiter/rhel6.git

They are in the btrfs-fixes branch.  I just git cherry picked all of them and
then any that I had to modify I put "FIXED FOR RHEL" in the subject and then a
list of the things I changed in the commit log.  IIRC there are only 3 or 4 of
them and they are mostly just one offs where we don't have an #include because
of some big tree-wide patch that I didn't take.  Thanks,

Josef

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>