thp+memcg-numa: fix BUG at mm.h:370
Patchwork-id: 34061
O-Subject: [RHEL6.1 PATCH] thp+memcg-numa: fix BUG at include/linux/mm.h:370!
Bugzilla: 687918
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=687918

This is just a precaution in case anybody builds the x86-64 kernels
with CONFIG_NUMA=n and memcg is enabled. AFIK all x86-64 kernels have
CONFIG_NUMA=y so this is really a noop for RHEL6/6.1. So it shouldn't
be urgent. Still it's safe enough that can go in 6.1 too I think.

The upstream fix is slightly different and it'll get more cleanups to
reduce the #ifdef but the raw fix (easier to review) is below.

====
Subject: thp+memcg-numa: fix BUG at include/linux/mm.h:370!

From: Hugh Dickins <hughd@google.com>

THP's collapse_huge_page() has an understandable but ugly difference
in when its huge page is allocated: inside if NUMA but outside if not.
It's hardly surprising that the memcg failure path forgot that, freeing
the page in the non-NUMA case, then hitting a VM_BUG_ON in get_page()
(or even worse, using the freed page).

Signed-off-by: Hugh Dickins <hughd@google.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>